backend/pyproject.toml
[project]
name = "llama-design-drive"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "annotated-types==0.7.0",
    "antlr4-python3-runtime==4.9.3",
    "anyio==4.6.0",
    "app[dev,test]",
    "asttokens==2.4.1",
    "certifi==2024.8.30",
    "colorama==0.4.6",
    "comm==0.2.2",
    "debugpy==1.8.1",
    "decorator==5.1.1",
    "distro==1.9.0",
    "exceptiongroup==1.2.0",
    "executing==2.0.1",
    "flask-cors>=5.0.0",
    "h11==0.14.0",
    "httpcore==1.0.5",
    "httpx==0.27.2",
    "hydra-core==1.3.2",
    "idna==3.10",
    "ipykernel==6.29.4",
    "ipython==8.23.0",
    "jedi==0.19.1",
    "jiter==0.5.0",
    "jupyter-client==8.6.1",
    "jupyter-core==5.7.2",
    "matplotlib-inline==0.1.7",
    "nest-asyncio==1.6.0",
    "ollama==0.3.3",
    "omegaconf==2.3.0",
    "openai==1.50.2",
    "packaging==24.0",
    "pandas>=2.2.3",
    "parso==0.8.4",
    "platformdirs==4.2.0",
    "prompt-toolkit==3.0.43",
    "psutil==5.9.8",
    "pure-eval==0.2.2",
    "pydantic-core==2.23.4",
    "pydantic==2.9.2",
    "pygments==2.17.2",
    "python-dateutil==2.9.0.post0",
    "pywin32==306",
    "pyyaml==6.0.2",
    "pyzmq==26.0.0",
    "six==1.16.0",
    "sniffio==1.3.1",
    "stack-data==0.6.3",
    "tornado==6.4",
    "tqdm==4.66.5",
    "traitlets==5.14.2",
    "typing-extensions>=4.12.2",
    "wcwidth==0.2.13",
]

[tool.uv.sources]
app = { path = "backend" }
